requirements.txt
Django==3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
PyJWT==1.7.1
python-decouple
wheel
psycopg2==2.8.6
PyJWT==1.7.1